pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "job-agent"
version = "0.0.1"
description = "Job search agent using Microsoft Agent Framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "agent-framework-azure-ai",
  "agent-framework-core",
  "azure-ai-agentserver-core",
  "azure-ai-agentserver-agentframework",
  "azure-identity",
  "python-dotenv",
  "asyncpg>=0.29.0",
  "pgvector>=0.3.0",
  "pymupdf>=1.24.0",
  "python-docx>=1.1.0",
  "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/job_agent"]
